AI-blog-backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
google-generativeai